+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Enhancements
 
+* [#2813](https://github.com/digitalocean/netbox/issues/2813) - Add tenant group filters
 * [#3150](https://github.com/digitalocean/netbox/issues/3150) - Fix formatting of cable length during cable trace
 * [#3085](https://github.com/digitalocean/netbox/issues/3085) - Catch all exceptions during export template rendering